draft version of argparser implementation
