+++ simtools/util/commandline_parser.py
@@ -1,5 +1,4 @@
 import argparse
-import logging
 
 
 class CommandLineParser(argparse.ArgumentParser):
@@ -15,12 +14,19 @@ class CommandLineParser(argparse.ArgumentParser):
         """
 
         self.add_argument(
+            "--test",
+            help="Test option for faster execution during development",
+            action="store_true",
+            required=False,
+        )
+        self.add_argument(
             "-v",
             "--verbosity",
             dest="logLevel",
             action="store",
             default="info",
             help="Log level to print (default is INFO)",
+            required=False,
         )
 
     def initialize_workflow_arguments(self):
@@ -58,3 +64,38 @@ class CommandLineParser(argparse.ArgumentParser):
         Initialize default arguments for telescope model definition
 
         """
+
+        self.add_argument(
+            "-s",
+            "--site",
+            help="North or South",
+            type=str,
+            required=True
+        )
+        self.add_argument(
+            "-t",
+            "--telescope",
+            help="Telescope model name (e.g. LST-1, SST-D, ...)",
+            type=str,
+            required=True,
+        )
+        self.add_argument(
+            "-m",
+            "--model_version",
+            help="Model version (default=Current)",
+            type=str,
+            default="Current",
+        )
+
+    @staticmethod
+    def efficiency_interval(value):
+        """
+        Argument parser check that value is an efficiency in the interval [0,1]
+
+        """
+        fvalue = float(value)
+        if fvalue < 0. or fvalue > 1.:
+            raise argparse.ArgumentTypeError(
+                "{} outside of allowed [0,1] interval".format(value))
+
+        return fvalue
